Made the ball change angles upon collision with the paddles based upon the placement of the hit. The angle is gradient from 0 to pi/3 radians.
Also made it so the paddles slowly resduce size (1 pixel per hit) . Note that it does not take into account pushing the ball. (suggestion of checking that the last time it reduced size was 1 second ago or more using the frames to tell).

Also fixed the bug that the ball was fast until its first hit.